pyproject.toml
[project]
name = "1141-hci-midterm"
version = "1.0.0"
description = "Stack Overflow Social Network Analysis (SNA) - HCI Midterm Project"
readme = "README.md"
authors = [
    { name = "TsukiSama9292", email = "a0985821880@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "gunicorn>=21.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "google-cloud-bigquery[db-dtypes]>=3.0.0",
    "pandas>=2.0.0",
    "igraph>=0.10.0",
    "matplotlib>=3.7.0",
    "db-dtypes>=1.5.1",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]
